+++ src/view/Inputs.js
