+++ src/view/Outputs.js
